Need to fix the Typeclass resolving
